# Talent-Hubzip/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jobera.az Job Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ratings:
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings for a candidate
      parameters:
        - name: candidateId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"

    post:
      operationId: createRating
      tags: [ratings]
      summary: HR submits rating and review for a candidate after interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, userId, name, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        employeeCount:
          type: ["string", "null"]
        isVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
...
    Rating:
      type: object
      required: [id, hrUserId, candidateId, stars, createdAt]
      properties:
        id:
          type: integer
        hrUserId:
          type: integer
        candidateId:
          type: integer
        company:
          $ref: "#/components/schemas/Company"
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RatingInput:
      type: object
      required: [candidateId, stars]
      properties:
        candidateId:
          type: integer
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...